machinecode: Freeze poison values for `assert_equal`
Make sure the `assert_equal` equality check doesn't get optimized away
because a poison value reaches it. This makes the check more robust. We
would never want for the assertion to pass because a poison value
reached it, and this lessens the chance of that happening, though it
doesn't rule it out.

Signed-off-by: Martin Povišer <povik@cutebit.org>