machinecode: Avoid passing assignments by value in IR
When implementing assignments, try to avoid having an IR value
corresponding to the user value being assigned, since those user values
can be arrays of great sizes, and this can lead to explosion of IR
instructions when a later compilation pass lowers operations to be per
element.

Instead emit `memcpy`, and an optional value-casting loop, for every
occurrence of an assignment. This can be changed back to passing by
value, or the `memcpy` eliminated by other means, by the LLVM passes
that follow later.

Signed-off-by: Martin Povišer <povik@cutebit.org>